cmd/search/job: Pull ProwJob and fetchJob into their own file
So it's more obviously independent of the main command logic.
